PPLUS-1506: Rework apply manifest method
