+++ src/ManifestStrategy/UnwireNavigationManifestStrategy.php
@@ -60,68 +60,28 @@ class UnwireNavigationManifestStrategy extends AbstractNavigationManifestStrateg
         $outputDiff = [];
 
         foreach ($navigation as $key => $value) {
-            if (array_key_exists($key, $manifestData)) {
-                if ($manifestData[$key] === null) {
-                    // Deleted element is found, do not add element to output array
-                } elseif (is_array($value) && is_array($manifestData[$key])) {
-                    $recursiveDiff = $this->applyNewNavigation($value, $manifestData[$key]);
-
-                    if (count($recursiveDiff)) {
-                        $outputDiff[$key] = $recursiveDiff;
-                    }
-                } elseif (!in_array($value, $manifestData)) {
+            if (!array_key_exists($key, $manifestData)) {
+                if (!in_array($value, $manifestData)) {
                     $outputDiff[$key] = $value;
                 }
-            } elseif (!in_array($value, $manifestData)) {
-                $outputDiff[$key] = $value;
+
+                continue;
+            }
+
+            if ($manifestData[$key] === null) {
+                // Deleted element is found, do not add element to output array
+                continue;
+            }
+
+            if (is_array($value) && is_array($manifestData[$key])) {
+                $recursiveDiff = $this->applyNewNavigation($value, $manifestData[$key]);
+
+                if (count($recursiveDiff)) {
+                    $outputDiff[$key] = $recursiveDiff;
+                }
             }
         }
 
         return $outputDiff;
-    }
-
-    /**
-     * @param array<string|int, array<int|string, mixed>> $navigation
-     * @param \SprykerSdk\Integrator\Dependency\Console\InputOutputInterface $inputOutput
-     * @param bool $isDry
-     *
-     * @return bool
-     */
-    protected function writeNavigationSchema(array $navigation, InputOutputInterface $inputOutput, bool $isDry): bool
-    {
-        $navigationXmlString = $this->getXmlNavigationFromArrayNavigation($navigation)->asXML();
-
-        if ($isDry) {
-            $inputOutput->writeln($navigationXmlString ?: '');
-
-            return true;
-        }
-
-        if ($navigationXmlString === false) {
-            return false;
-        }
-
-        $navigationXmlDomDocument = new DOMDocument('1.0');
-        $navigationXmlDomDocument->preserveWhiteSpace = false;
-        $navigationXmlDomDocument->formatOutput = true;
-        $navigationXmlDomDocument->loadXML($navigationXmlString);
-
-        $callback = function ($matches) {
-            $multiplier = (int)(strlen($matches[1]) / 2) * 4;
-
-            return str_repeat(' ', $multiplier) . '<';
-        };
-
-        $navigationXmlString = $navigationXmlDomDocument->saveXML();
-
-        if ($navigationXmlString === false) {
-            return false;
-        }
-
-        $content = preg_replace_callback('/^( +)</m', $callback, $navigationXmlString);
-
-        file_put_contents($this->getNavigationSourceFilePath(), $content);
-
-        return true;
     }
 }
